# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Velvet Store API - Premium adult wellness ecommerce for Kenya
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders, admin]
      summary: List orders
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, processing, out_for_delivery, delivered, cancelled]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"

    post:
      operationId: createOrder
      tags: [orders]
      summary: Create an order (checkout)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderBody"
      responses:
        "201":
          description: Created order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

    put:
      operationId: updateOrder
      tags: [orders, admin]
      summary: Update order status / notes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderBody"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"

  # ── ADMIN DASHBOARD ────────────────────────────────────────────
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        status:
          type: string
          enum: [pending, confirmed, processing, out_for_delivery, delivered, cancelled]
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        deliveryAddress:
          type: string
        deliveryRegion:
          type: string
          enum: [nairobi, outside_nairobi]
        shippingFee:
          type: number
        subtotal:
          type: number
        discountAmount:
          type: number
        total:
          type: number
        couponCode:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        internalNotes:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [cash_on_delivery, manual_payment, pending]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, orderNumber, status, customerPhone, isAnonymous, deliveryAddress, deliveryRegion, shippingFee, subtotal, discountAmount, total, paymentMethod, items, createdAt, updatedAt]
    OrderItem:
      type: object
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        productTitle:
          type: string
        productImageUrl:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required: [id, orderId, productId, productTitle, quantity, unitPrice, totalPrice]
    OrderListResponse:
      type: object
      properties:
        orders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        totalPages:
          type: integer
      required: [orders, total, page, totalPages]
    CreateOrderBody:
      type: object
      properties:
        customerName:
          type: string
          nullable: true
        customerPhone:
          type: string
        customerEmail:
          type: string
          nullable: true
        isAnonymous:
          type: boolean
        deliveryAddress:
          type: string
        deliveryRegion:
          type: string
          enum: [nairobi, outside_nairobi]
        notes:
          type: string
          nullable: true
        couponCode:
          type: string
          nullable: true
        paymentMethod:
          type: string
          enum: [cash_on_delivery, manual_payment, pending]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"
      required: [customerPhone, deliveryAddress, deliveryRegion, paymentMethod, items]
    OrderItemInput:
      type: object
      properties:
        productId:
          type: integer
        quantity:
          type: integer
      required: [productId, quantity]
    UpdateOrderBody:
      type: object
      properties:
        status:
          type: string
          enum: [pending, confirmed, processing, out_for_delivery, delivered, cancelled]
        internalNotes:
          type: string
          nullable: true

    # ── ADMIN SCHEMAS ───────────────────────────────────────────
...